+++ changelog.md
@@ -1,3 +1,8 @@
+# 1.0.2
+
+-   Now it's possible to pass the lambda function from dictionary file.
+-   with_confirmation parameter added.
+
 # 1.0.1
 
 -   Strings with image links are ignored.